# refactor:  enhance readability and performance of graph.adjacency.sparse