--- a/R/adjacency.R
+++ b/R/adjacency.R
@@ -1,4 +1,3 @@
-
 #' Create graphs from adjacency matrices
 #'
 #' @description
@@ -119,15 +118,15 @@ graph.adjacency <- function(adjmatrix, mode = c("directed", "undirected", "max",
 #' @examples
 #'
 #' g1 <- sample(
-#'     x = 0:1, size = 100, replace = TRUE,
-#'     prob = c(0.9, 0.1)
-#'   ) %>%
+#'   x = 0:1, size = 100, replace = TRUE,
+#'   prob = c(0.9, 0.1)
+#' ) %>%
 #'   matrix(ncol = 10) %>%
 #'   graph_from_adjacency_matrix()
 #'
 #' g2 <- sample(
-#'     x = 0:5, size = 100, replace = TRUE,
-#'     prob = c(0.9, 0.02, 0.02, 0.02, 0.02, 0.02)
+#'   x = 0:5, size = 100, replace = TRUE,
+#'   prob = c(0.9, 0.02, 0.02, 0.02, 0.02, 0.02)
 #' ) %>%
 #'   matrix(ncol = 10) %>%
 #'   graph_from_adjacency_matrix(weighted = TRUE)
@@ -193,8 +192,8 @@ graph.adjacency <- function(adjmatrix, mode = c("directed", "undirected", "max",
 #'   x
 #' }
 #' expected_g8_weights <- non_zero_sort(
-#'   halve_diag(adj_matrix + t(adj_matrix)
-#' )[lower.tri(adj_matrix, diag = TRUE)])
+#'   halve_diag(adj_matrix + t(adj_matrix))[lower.tri(adj_matrix, diag = TRUE)]
+#' )
 #' actual_g8_weights <- sort(E(g8)$weight)
 #' all(expected_g8_weights == actual_g8_weights)
 #'
@@ -229,7 +228,6 @@ graph_from_adjacency_matrix <- function(adjmatrix,
                                         ),
                                         weighted = NULL, diag = TRUE,
                                         add.colnames = NULL, add.rownames = NA) {
-
   mode <- igraph.match.arg(mode)
 
   if (!is.matrix(adjmatrix) && !inherits(adjmatrix, "Matrix")) {
@@ -378,156 +376,55 @@ mysummary <- function(x) {
 }
 
 
-graph.adjacency.sparse <- function(adjmatrix, mode, weighted = NULL, diag = TRUE) {
+graph.adjacency.sparse <- function(adjmatrix, mode, weighted = NULL, diag = TRUE, call = rlang::caller_env()) {
   if (!is.null(weighted)) {
     if (is.logical(weighted) && weighted) {
       weighted <- "weight"
     }
     if (!is.character(weighted)) {
-      stop("invalid value supplied for `weighted' argument, please see docs.")
+      cli::cli_abort("Invalid value supplied for `weighted' argument, please see docs.", call = call)
     }
   }
 
   if (nrow(adjmatrix) != ncol(adjmatrix)) {
-    stop("not a square matrix")
+    cli::cli_abort("Not a square matrix", call = call)
   }
 
   vc <- nrow(adjmatrix)
+  # Exit early for empty graphs
+  if (vc == 1 || Matrix::nnzero(adjmatrix) == 0) {
+    return(make_empty_graph(n = vc, directed = (mode == "directed")))
+  }
 
-  ## to remove non-redundancies that can persist in a dgtMatrix
   if (inherits(adjmatrix, "dgTMatrix")) {
     adjmatrix <- as(adjmatrix, "CsparseMatrix")
   } else if (inherits(adjmatrix, "ddiMatrix")) {
     adjmatrix <- as(adjmatrix, "CsparseMatrix")
   }
 
-  if (mode == "directed") {
-    ## DIRECTED
-    el <- mysummary(adjmatrix)
-    if (!diag) {
-      el <- el[el[, 1] != el[, 2], ]
-    }
-  } else if (mode == "undirected") {
-    ## UNDIRECTED, must be symmetric if weighted
+  if (mode == "undirected") {
     if (!is.null(weighted) && !Matrix::isSymmetric(adjmatrix)) {
       stop("Please supply a symmetric matrix if you want to create a weighted graph with mode=UNDIRECTED.")
     }
-    if (diag) {
-      adjmatrix <- Matrix::tril(adjmatrix)
-    } else {
-      if (vc == 1) {
-        # Work around Matrix glitch
-        adjmatrix <- as(matrix(0), "dgCMatrix")
-      } else {
-        adjmatrix <- Matrix::tril(adjmatrix, -1)
-      }
-    }
-    el <- mysummary(adjmatrix)
-    rm(adjmatrix)
-  } else if (mode == "max") {
-    ## MAXIMUM
-    el <- mysummary(adjmatrix)
-    rm(adjmatrix)
-    if (!diag) {
-      el <- el[el[, 1] != el[, 2], ]
-    }
-    el <- el[el[, 3] != 0, ]
-    w <- el[, 3]
-    el <- el[, 1:2]
-    el <- cbind(pmin(el[, 1], el[, 2]), pmax(el[, 1], el[, 2]))
-    o <- order(el[, 1], el[, 2])
-    el <- el[o, , drop = FALSE]
-    w <- w[o]
-    if (nrow(el) > 1) {
-      dd <- el[2:nrow(el), 1] == el[1:(nrow(el) - 1), 1] &
-        el[2:nrow(el), 2] == el[1:(nrow(el) - 1), 2]
-      dd <- which(dd)
-      if (length(dd) > 0) {
-        mw <- pmax(w[dd], w[dd + 1])
-        w[dd] <- mw
-        w[dd + 1] <- mw
-        el <- el[-dd, , drop = FALSE]
-        w <- w[-dd]
-      }
-    }
-    el <- cbind(el, w)
-  } else if (mode == "upper") {
-    ## UPPER
-    if (diag) {
-      adjmatrix <- Matrix::triu(adjmatrix)
-    } else {
-      adjmatrix <- Matrix::triu(adjmatrix, 1)
-    }
-    el <- mysummary(adjmatrix)
-    rm(adjmatrix)
-    if (!diag) {
-      el <- el[el[, 1] != el[, 2], ]
-    }
+    adjmatrix <- Matrix::tril(adjmatrix)
   } else if (mode == "lower") {
-    ## LOWER
-    if (diag) {
-      adjmatrix <- Matrix::tril(adjmatrix)
-    } else {
-      if (vc == 1) {
-        # Work around Matrix glitch
-        adjmatrix <- as(matrix(0), "dgCMatrix")
-      } else {
-        adjmatrix <- Matrix::tril(adjmatrix, -1)
-      }
-    }
-    el <- mysummary(adjmatrix)
-    rm(adjmatrix)
-    if (!diag) {
-      el <- el[el[, 1] != el[, 2], ]
-    }
-  } else if (mode == "min") {
-    ## MINIMUM
-    adjmatrix <- sign(adjmatrix) * sign(Matrix::t(adjmatrix)) * adjmatrix
-    el <- mysummary(adjmatrix)
-    rm(adjmatrix)
-    if (!diag) {
-      el <- el[el[, 1] != el[, 2], ]
-    }
-    el <- el[el[, 3] != 0, ]
-    w <- el[, 3]
-    el <- el[, 1:2]
-    el <- cbind(pmin(el[, 1], el[, 2]), pmax(el[, 1], el[, 2]))
-    o <- order(el[, 1], el[, 2])
-    el <- el[o, ]
-    w <- w[o]
-    if (nrow(el) > 1) {
-      dd <- el[2:nrow(el), 1] == el[1:(nrow(el) - 1), 1] &
-        el[2:nrow(el), 2] == el[1:(nrow(el) - 1), 2]
-      dd <- which(dd)
-      if (length(dd) > 0) {
-        mw <- pmin(w[dd], w[dd + 1])
-        w[dd] <- mw
-        w[dd + 1] <- mw
-        el <- el[-dd, ]
-        w <- w[-dd]
-      }
-    }
-    el <- cbind(el, w)
+    adjmatrix <- Matrix::tril(adjmatrix)
+  } else if (mode == "upper") {
+    adjmatrix <- Matrix::triu(adjmatrix)
   } else if (mode == "plus") {
-    ## PLUS
     adjmatrix <- adjmatrix + Matrix::t(adjmatrix)
-    if (diag) {
-      adjmatrix <- Matrix::tril(adjmatrix)
-    } else {
-      if (vc == 1) {
-        # Work around Matrix glitch
-        adjmatrix <- as(matrix(0), "dgCMatrix")
-      } else {
-        adjmatrix <- Matrix::tril(adjmatrix, -1)
-      }
-    }
-    el <- mysummary(adjmatrix)
-    rm(adjmatrix)
-    if (diag) {
-      loop <- el[, 1] == el[, 2]
-      el[loop, 3] <- el[loop, 3] / 2
-    }
-    el <- el[el[, 3] != 0, ]
+    adjmatrix <- Matrix::tril(adjmatrix)
+  } else if (mode == "max") {
+    adjmatrix <- pmax(adjmatrix, Matrix::t(adjmatrix))
+    adjmatrix <- Matrix::tril(adjmatrix)
+  } else if (mode == "min") {
+    adjmatrix <- pmin(adjmatrix, Matrix::t(adjmatrix))
+    adjmatrix <- Matrix::tril(adjmatrix)
+    adjmatrix <- Matrix::drop0(adjmatrix)
+  }
+  el <- mysummary(adjmatrix)
+  if (!diag) {
+    el <- el[el[, 1] != el[, 2]]
   }
 
   if (!is.null(weighted)) {
@@ -536,7 +433,12 @@ graph.adjacency.sparse <- function(adjmatrix, mode, weighted = NULL, diag = TRUE
     names(weight) <- weighted
     res <- add_edges(res, edges = t(as.matrix(el[, 1:2])), attr = weight)
   } else {
-    edges <- unlist(apply(el, 1, function(x) rep(unname(x[1:2]), x[3])))
+    if (max(el[, 3]) == 1) {
+      edges <- as.vector(t(el[, 1:2]))
+    } else {
+      row_repeats <- rep(seq_len(nrow(el)), times = el[, 3])
+      edges <- as.vector(t(el[row_repeats, 1:2]))
+    }
     res <- make_graph(n = vc, edges, directed = (mode == "directed"))
   }
   res