This PR is a large refactoring of `graph.adjacency.sparse()`. The goal was to make the function more readable and remove some performance bottlenecks. 

The only cases that has a slightly worse performance in some cases are `mode = "min"` and `mode = "max"` (However, there are some fluctuations I do not fully understand). The new implementation is a simple one-liner. Probably needs some discussions.

``` r
library(igraph)
packageVersion("igraph")
#> [1] '2.1.3'

A <- sample_gnp(5000,0.02, directed = FALSE) |> as_adjacency_matrix()

bench::mark(
  check = FALSE,
  old = igraph:::graph.adjacency.sparse(A,mode = "undirected"),
  new = graph.adjacency.sparse.new(A,mode = "undirected")
)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 2 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 old        515.15ms  515.2ms      1.94    32.6MB     25.2
#> 2 new          8.69ms   10.1ms     92.7       20MB     39.4

A[A==1] <- sample(1:10,sum(A==1),replace = TRUE)

bench::mark(
  check = FALSE,
  old = igraph:::graph.adjacency.sparse(A,mode = "max"),
  new = graph.adjacency.sparse.new(A,mode = "max")
)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 2 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 old           586ms    586ms      1.71     198MB    13.7 
#> 2 new           783ms    783ms      1.28     645MB     3.83

bench::mark(
  check = FALSE,
  old = igraph:::graph.adjacency.sparse(A,mode = "max", weighted = TRUE),
  new = graph.adjacency.sparse.new(A,mode = "max", weighted = TRUE)
)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 2 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 old          43.7ms   54.4ms     18.4      150MB     9.20
#> 2 new         128.9ms  184.3ms      5.40     473MB    14.4

A <- sample_gnp(5000,0.02, directed = TRUE) |> as_adjacency_matrix()

bench::mark(
  check = FALSE,
  old = igraph:::graph.adjacency.sparse(A,mode = "directed"),
  new = graph.adjacency.sparse.new(A,mode = "directed")
)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 2 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 old           1.09s    1.09s     0.918    59.2MB    23.9 
#> 2 new         18.58ms  19.66ms    47.3      24.8MB     5.91
```

<sup>Created on 2025-01-28 with [reprex v2.1.1](https://reprex.tidyverse.org)</sup>